Move tests to ToolsTests
