+++ src/Cake.Common/Tools/DotNet/DotNetSettings.cs
@@ -2,8 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
-using Cake.Core.IO;
 using Cake.Core.Tooling;
 
 namespace Cake.Common.Tools.DotNet
